# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Hevy API Docs",
        "description": "\n      Welcome to Hevy's public API! We're just starting to roll this out and depending on your\n      feedback, we'll be adding more features and endpoints. Also, we make no guarantees that \n      we won't completely change the structure or abandon the project entirely so use it at\n      your own risk. Currently, this API is only available to Hevy Pro users. You can get your \n      key on our web app at https://hevy.com/settings?developer. If you have any questions,\n      please reach out to pedro@hevyapp.com 🤝.\n      ",
        "version": "0.0.1"
    },
    "paths": {
...
        "/v1/exercise_templates": {
            "get": {
                "summary": "Get a paginated list of exercise templates available on the account.",
                "tags": [
                    "ExerciseTemplates"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number (Must be 1 or greater)"
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        },
                        "description": "Number of items on the requested page (Max 100)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of exercise templates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "page": {
                                            "type": "integer",
                                            "default": 1,
                                            "description": "Current page number"
                                        },
                                        "page_count": {
                                            "type": "integer",
                                            "default": 5,
                                            "description": "Total number of pages"
                                        },
                                        "exercise_templates": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ExerciseTemplate"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid page size"
                    }
                }
            },
            "post": {
                "summary": "Create a new custom exercise template.",
                "tags": [
                    "ExerciseTemplates"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCustomExerciseRequestBody"
                            }
                        }
                    },
                    "description": "The exercise template to create."
                },
                "responses": {
                    "200": {
                        "description": "The exercise template was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "description": "The ID of the exercise template",
                                            "example": 123
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message",
                                            "example": "Invalid request body"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Exceeds custom exercise limit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message",
                                            "example": "exceeds-custom-exercise-limit"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/exercise_templates/{exerciseTemplateId}": {
            "get": {
                "tags": [
                    "ExerciseTemplates"
                ],
                "summary": "Get a single exercise template by id.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "name": "exerciseTemplateId",
                        "in": "path",
                        "description": "The id of the exercise template",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExerciseTemplate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercise template not found"
                    }
                }
            }
        },
...
        "/v1/body_measurements": {
            "get": {
                "summary": "Get a paginated list of body measurements for the authenticated user",
                "tags": [
                    "Measurements"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number (Must be 1 or greater)"
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        },
                        "description": "Number of items per page (Max 10)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of body measurements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "page_count": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "body_measurements": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BodyMeasurement"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid page or pageSize"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            },
            "post": {
                "summary": "Create a body measurement entry for a given date. Returns 409 if an entry already exists for that date.",
                "tags": [
                    "Measurements"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BodyMeasurement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The measurement was successfully created"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "A measurement for this date already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/body_measurements/{date}": {
            "get": {
                "summary": "Get a single body measurement by date",
                "tags": [
                    "Measurements"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-15"
                        },
                        "description": "The date of the body measurement (YYYY-MM-DD)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The body measurement for the given date",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BodyMeasurement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Body measurement not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update an existing body measurement entry for a given date. All fields are overwritten; omitted fields are set to null.",
                "tags": [
                    "Measurements"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "api-key",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-15"
                        },
                        "description": "The date of the measurement to update (YYYY-MM-DD)"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutBodyMeasurement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The measurement was successfully updated"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No measurement found for the given date",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "BodyMeasurement": {
                "type": "object",
                "required": [
                    "date"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-14"
                    },
                    "weight_kg": {
                        "type": "number",
                        "nullable": true,
                        "example": 80.5
                    },
                    "lean_mass_kg": {
                        "type": "number",
                        "nullable": true,
                        "example": 65
                    },
                    "fat_percent": {
                        "type": "number",
                        "nullable": true,
                        "example": 18.5
                    },
                    "neck_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 38
                    },
                    "shoulder_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 115
                    },
                    "chest_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 95
                    },
                    "left_bicep_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 35
                    },
                    "right_bicep_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 35.5
                    },
                    "left_forearm_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 28
                    },
                    "right_forearm_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 28.5
                    },
                    "abdomen": {
                        "type": "number",
                        "nullable": true,
                        "example": 85
                    },
                    "waist": {
                        "type": "number",
                        "nullable": true,
                        "example": 80
                    },
                    "hips": {
                        "type": "number",
                        "nullable": true,
                        "example": 95
                    },
                    "left_thigh": {
                        "type": "number",
                        "nullable": true,
                        "example": 55
                    },
                    "right_thigh": {
                        "type": "number",
                        "nullable": true,
                        "example": 55.5
                    },
                    "left_calf": {
                        "type": "number",
                        "nullable": true,
                        "example": 37
                    },
                    "right_calf": {
                        "type": "number",
                        "nullable": true,
                        "example": 37.5
                    }
                }
            },
            "PutBodyMeasurement": {
                "type": "object",
                "properties": {
                    "weight_kg": {
                        "type": "number",
                        "nullable": true,
                        "example": 80.5
                    },
                    "lean_mass_kg": {
                        "type": "number",
                        "nullable": true,
                        "example": 65
                    },
                    "fat_percent": {
                        "type": "number",
                        "nullable": true,
                        "example": 18.5
                    },
                    "neck_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 38
                    },
                    "shoulder_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 115
                    },
                    "chest_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 95
                    },
                    "left_bicep_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 35
                    },
                    "right_bicep_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 35.5
                    },
                    "left_forearm_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 28
                    },
                    "right_forearm_cm": {
                        "type": "number",
                        "nullable": true,
                        "example": 28.5
                    },
                    "abdomen": {
                        "type": "number",
                        "nullable": true,
                        "example": 85
                    },
                    "waist": {
                        "type": "number",
                        "nullable": true,
                        "example": 80
                    },
                    "hips": {
                        "type": "number",
                        "nullable": true,
                        "example": 95
                    },
                    "left_thigh": {
                        "type": "number",
                        "nullable": true,
                        "example": 55
                    },
                    "right_thigh": {
                        "type": "number",
                        "nullable": true,
                        "example": 55.5
                    },
                    "left_calf": {
                        "type": "number",
                        "nullable": true,
                        "example": 37
                    },
                    "right_calf": {
                        "type": "number",
                        "nullable": true,
                        "example": 37.5
                    }
                }
            },
...
            "CustomExerciseType": {
                "type": "enum",
                "enum": [
                    "weight_reps",
                    "reps_only",
                    "bodyweight_reps",
                    "bodyweight_assisted_reps",
                    "duration",
                    "weight_duration",
                    "distance_duration",
                    "short_distance_weight"
                ],
                "example": "weight_reps"
            },
            "MuscleGroup": {
                "type": "enum",
                "enum": [
                    "abdominals",
                    "shoulders",
                    "biceps",
                    "triceps",
                    "forearms",
                    "quadriceps",
                    "hamstrings",
                    "calves",
                    "glutes",
                    "abductors",
                    "adductors",
                    "lats",
                    "upper_back",
                    "traps",
                    "lower_back",
                    "chest",
                    "cardio",
                    "neck",
                    "full_body",
                    "other"
                ],
                "example": "chest"
            },
            "EquipmentCategory": {
                "type": "enum",
                "enum": [
                    "none",
                    "barbell",
                    "dumbbell",
                    "kettlebell",
                    "machine",
                    "plate",
                    "resistance_band",
                    "suspension",
                    "other"
                ],
                "example": "barbell"
            },
            "ExerciseTemplate": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "The exercise template ID.",
                        "example": "b459cba5-cd6d-463c-abd6-54f8eafcadcb"
                    },
                    "title": {
                        "type": "string",
                        "description": "The exercise title.",
                        "example": "Bench Press (Barbell)"
                    },
                    "type": {
                        "type": "string",
                        "description": "The exercise type.",
                        "example": "weight_reps"
                    },
                    "primary_muscle_group": {
                        "type": "string",
                        "description": "The primary muscle group of the exercise.",
                        "example": "chest"
                    },
                    "secondary_muscle_groups": {
                        "type": "array",
                        "description": "The secondary muscle groups of the exercise.",
                        "items": {
                            "type": "string"
                        }
                    },
                    "is_custom": {
                        "type": "boolean",
                        "description": "A boolean indicating whether the exercise is a custom exercise.",
                        "example": false
                    }
                }
            },
            "CreateCustomExerciseRequestBody": {
                "type": "object",
                "properties": {
                    "exercise": {
                        "required": true,
                        "type": "object",
                        "properties": {
                            "title": {
                                "type": "string",
                                "description": "The title of the exercise template.",
                                "example": "Bench Press"
                            },
                            "exercise_type": {
                                "$ref": "#/components/schemas/CustomExerciseType"
                            },
                            "equipment_category": {
                                "$ref": "#/components/schemas/EquipmentCategory",
                                "description": "The equipment category of the exercise template.",
                                "example": "barbell"
                            },
                            "muscle_group": {
                                "$ref": "#/components/schemas/MuscleGroup",
                                "description": "The muscle group of the exercise template.",
                                "example": "chest"
                            },
                            "other_muscles": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/MuscleGroup"
                                },
                                "description": "The other muscles of the exercise template.",
                                "example": [
                                    "biceps",
                                    "triceps"
                                ]
                            }
                        }
                    }
                }
            },
...
        }
    },
...
}